refactor: separate NDD observability
